create computer random word game
Create a game where the computer picks a random word and
the player has to guess that word. The computer tells the
player how many letters are in the word. Then the player gets
five chances to ask if a letter is in the word. The computer can
only respond with “yes” or “no”. Then, the player must guess
the word.